+++ DatadogCore/Tests/DatadogObjc/ObjcAPITests/DDSessionReplay+apiTests.m
@@ -21,4 +21,14 @@
     [DDSessionReplay enableWith:configuration];
 }
 
+- (void)testConfigurationWithNewApi {
+    DDSessionReplayConfiguration *configuration = [[DDSessionReplayConfiguration alloc] initWithReplaySampleRate:100
+                                                   textAndInputPrivacyLevel:DDTextAndInputPrivacyLevelMaskAll
+                                                   imagePrivacyLevel:DDImagePrivacyLevelMaskNone
+                                                   touchPrivacyLevel:DDTouchPrivacyLevelShow];
+    configuration.customEndpoint = [NSURL new];
+
+    [DDSessionReplay enableWith:configuration];
+}
+
 @end