+++ DatadogRUM/Sources/Integrations/TelemetryReceiver.swift
@@ -80,12 +80,20 @@ internal final class TelemetryReceiver: FeatureMessageReceiver {
         case .configuration(let configuration):
             send(configuration: configuration)
         case let .metric(metric):
-            send(metric: metric)
+            if sampled(event: metric) {
+                send(metric: metric)
+            }
         case .usage(let usage):
-            send(usage: usage)
+            if sampled(event: usage) {
+                send(usage: usage)
+            }
         }
 
         return true
+    }
+
+    private func sampled(event: SampledTelemetry) -> Bool {
+        return Sampler(samplingRate: event.sampleRate).sample()
     }
 
     /// Sends a `TelemetryDebugEvent` event.
@@ -237,10 +245,6 @@ internal final class TelemetryReceiver: FeatureMessageReceiver {
     }
 
     private func send(metric: MetricTelemetry) {
-        guard Sampler(samplingRate: metric.sampleRate).sample() else {
-            return
-        }
-
         let date = dateProvider.now
 
         record(event: nil) { context, writer in
@@ -306,7 +310,7 @@ internal final class TelemetryReceiver: FeatureMessageReceiver {
 
 private extension TelemetryUsageEvent.Telemetry.Usage {
     init(_ usage: UsageTelemetry) {
-        switch usage {
+        switch usage.event {
         case .setTrackingConsent(let consent):
             self = .telemetryCommonFeaturesUsage(value: .setTrackingConsent(value: .init(trackingConsent: .init(consent: consent))))
         case .stopSession:
@@ -367,6 +371,7 @@ private extension TelemetryConfigurationEvent.Telemetry.Configuration {
             forwardConsoleLogs: nil,
             forwardErrorsToLogs: nil,
             forwardReports: nil,
+            imagePrivacyLevel: configuration.imagePrivacyLevel,
             initializationType: nil,
             mobileVitalsUpdatePeriod: configuration.mobileVitalsUpdatePeriod,
             premiumSampleRate: nil,
@@ -374,7 +379,7 @@ private extension TelemetryConfigurationEvent.Telemetry.Configuration {
             reactVersion: nil,
             replaySampleRate: nil,
             selectedTracingPropagators: nil,
-            sessionReplaySampleRate: nil,
+            sessionReplaySampleRate: configuration.sessionReplaySampleRate,
             sessionSampleRate: configuration.sessionSampleRate,
             silentMultipleInit: nil,
             startRecordingImmediately: configuration.startRecordingImmediately,
@@ -382,6 +387,8 @@ private extension TelemetryConfigurationEvent.Telemetry.Configuration {
             telemetryConfigurationSampleRate: nil,
             telemetrySampleRate: configuration.telemetrySampleRate,
             telemetryUsageSampleRate: nil,
+            textAndInputPrivacyLevel: configuration.textAndInputPrivacyLevel,
+            touchPrivacyLevel: configuration.touchPrivacyLevel,
             traceSampleRate: configuration.traceSampleRate,
             tracerApi: configuration.tracerAPI,
             tracerApiVersion: configuration.tracerAPIVersion,