+++ IntegrationTests/Runner/Scenarios/SessionReplay/SRScenarios.swift
@@ -24,6 +24,7 @@ final class SRMultipleViewsRecordingScenario: TestScenario {
 
         var srConfig = SessionReplay.Configuration(replaySampleRate: 100)
         srConfig.defaultPrivacyLevel = .mask
+        srConfig.touchPrivacyLevel = .show
         srConfig.customEndpoint = Environment.serverMockConfiguration()?.srEndpoint
         SessionReplay.enable(with: srConfig)
     }