+++ CHANGELOG.md
@@ -1,6 +1,13 @@
+# Unreleased
+
+# 2.18.0 / 25-09-2024
+- [IMPROVEMENT] Add overwrite required (breaking) param to addViewLoadingTime & usage telemetry. See [#2040][]
+- [FEATURE] Prevent "show password" features from revealing sensitive texts in Session Replay. See [#2050][]
+- [FEATURE] Add Fine-Grained Masking configuration options to Session Replay. See [#2043][]
+
 # 2.17.0 / 11-09-2024
 
-- [FEATURE] Add support for view loading API (addViewLoadingTime). See [#2026][]
+- [FEATURE] Add support for view loading experimental API (addViewLoadingTime). See [#2026][]
 - [IMPROVEMENT] Drop support for deprecated cocoapod specs. See [#1998][]
 - [FIX] Propagate global Tracer tags to OpenTelemetry span attributes. See [#2000][]
 - [FEATURE] Add Logs event mapper to ObjC API. See [#2008][]
@@ -764,6 +771,9 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#1998]: https://github.com/DataDog/dd-sdk-ios/pull/1998
 [#1966]: https://github.com/DataDog/dd-sdk-ios/pull/1966
 [#2026]: https://github.com/DataDog/dd-sdk-ios/pull/2026
+[#2043]: https://github.com/DataDog/dd-sdk-ios/pull/2043
+[#2040]: https://github.com/DataDog/dd-sdk-ios/pull/2040
+[#2050]: https://github.com/DataDog/dd-sdk-ios/pull/2050
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin
 [@hengyu]: https://github.com/Hengyu